Add Phase 1 runner and fix torch_dtype deprecation
The runner iterates every YAML task under the phase1 benchmark
directory and writes one JSONL trajectory per task under
runs/<model_slug>/<task_id>.jsonl, so a single command reproduces the
full Phase 1 data pull for a given target model. Also renames the
from_pretrained kwarg from the deprecated torch_dtype to dtype,
silencing the transformers >=5 warning.

+++ src/agent/agent.py
@@ -114,9 +114,12 @@ class Agent:
         self._tokenizer = AutoTokenizer.from_pretrained(
             model_name, trust_remote_code=trust_remote_code
         )
+        # transformers >=5 renamed `torch_dtype` to `dtype`; the old
+        # kwarg still works but emits a deprecation warning, so pass
+        # the new name explicitly.
         self._model = AutoModelForCausalLM.from_pretrained(
             model_name,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             trust_remote_code=trust_remote_code,
         ).to(device)
         self._model.eval()
